improv/media-code-sharing: refactor core logic to separate module with no react deps.
